+++ packages/types/src/interfaces/staking/definitions.ts
@@ -2,8 +2,17 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+const deprecated = {
+  Points: 'u32',
+  EraPoints: {
+    total: 'Points',
+    individual: 'Vec<Points>'
+  }
+};
+
 export default {
   types: {
+    ...deprecated,
     CompactAssignments: {
       votes1: 'Vec<(AccountId, [CompactScore; 0], AccountId)>',
       votes2: 'Vec<(AccountId, [CompactScore; 1], AccountId)>',
@@ -39,9 +48,9 @@ export default {
       }
     },
     EraIndex: 'u32',
-    EraPoints: {
-      total: 'Points',
-      individual: 'Vec<Points>'
+    EraRewardPoints: {
+      total: 'RewardPoint',
+      individual: 'BTreeMap<AccountId, RewardPoint>'
     },
     EraRewards: {
       total: 'u32',
@@ -80,6 +89,7 @@ export default {
         'Controller'
       ]
     },
+    RewardPoint: 'u32',
     SlashJournalEntry: {
       who: 'AccountId',
       amount: 'Balance',
@@ -101,12 +111,21 @@ export default {
       slashed: 'Balance',
       paidOut: 'Balance'
     },
-    StakingLedger: {
+    StakingLedgerTo223: {
       stash: 'AccountId',
       total: 'Compact<Balance>',
       active: 'Compact<Balance>',
       unlocking: 'Vec<UnlockChunk>'
     },
+    // TODO Enable as default when new staking payouts go live
+    StakingLedgerNew: {
+      stash: 'AccountId',
+      total: 'Compact<Balance>',
+      active: 'Compact<Balance>',
+      unlocking: 'Vec<UnlockChunk>',
+      nextReward: 'EraIndex'
+    },
+    StakingLedger: 'StakingLedgerTo223',
     UnappliedSlashOther: '(AccountId, Balance)',
     UnappliedSlash: {
       validator: 'AccountId',